astminer compatibility separated as package
